Refactoring generator headers to include radial/s2 generators
